apollo server installed as well as other bugs fixed
