+++ client/src/components/Chat.jsx
@@ -1,6 +1,6 @@
 import io from "socket.io-client";
 import { useEffect, useState } from "react";
-const socket = io.connect("http://localhost:3002");
+const socket = io.connect("http://localhost:3000");
 function Chat() {
   const [room, setRoom] = useState("");
   const [message, setMessage] = useState("");
