+++ client/src/components/Login.jsx
@@ -1,25 +1,76 @@
-import Button from "react-bootstrap/Button";
-import Form from "react-bootstrap/Form";
+import React, { useState } from "react";
+import { useMutation } from "@apollo/client";
+import { LOGIN_USER } from "../utils/mutation"; 
+const Login = () => {
+  const [formData, setFormData] = useState({
+    username: "",
+    password: "",
+  });
 
-export default function Login() {
+  const [loginUser, { loading, error }] = useMutation(LOGIN_USER);
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    const { username, password } = formData;
+
+    // Call the loginUser mutation with the form data
+    loginUser({
+      variables: {
+        username,
+        password,
+      },
+    })
+      .then((response) => {
+        console.log("User logged in successfully:", response.data);
+        // Optionally, redirect the user to the dashboard or another protected page
+        // history.push("/dashboard");
+      })
+      .catch((error) => {
+        console.error("User login failed:", error);
+      });
+  };
+
+  const handleChange = (e) => {
+    const { name, value } = e.target;
+    setFormData((prevData) => ({
+      ...prevData,
+      [name]: value,
+    }));
+  };
+
   return (
-    <Form>
-      <Form.Group className="mb-3" controlId="formBasicEmail">
-        <Form.Label>Email address</Form.Label>
-        <Form.Control type="email" placeholder="Enter email" />
-        <Form.Text className="text-muted">
-          We'll never share your email with anyone else.
-        </Form.Text>
-      </Form.Group>
-
-      <Form.Group className="mb-3" controlId="formBasicPassword">
-        <Form.Label>Password</Form.Label>
-        <Form.Control type="password" placeholder="Password" />
-      </Form.Group>
-
-      <Button variant="primary" type="submit">
-        Submit
-      </Button>
-    </Form>
+    <div>
+      <h2>Login</h2>
+      <form onSubmit={handleSubmit}>
+        <div>
+          <label htmlFor="username">Username:</label>
+          <input
+            type="text"
+            id="username"
+            name="username"
+            value={formData.username}
+            onChange={handleChange}
+            required
+          />
+        </div>
+        <div>
+          <label htmlFor="password">Password:</label>
+          <input
+            type="password"
+            id="password"
+            name="password"
+            value={formData.password}
+            onChange={handleChange}
+            required
+          />
+        </div>
+        <button type="submit" disabled={loading}>
+          Login
+        </button>
+        {error && <p>Error: {error.message}</p>}
+      </form>
+    </div>
   );
-}
+};
+
+export default Login;
