+++ package.json
@@ -16,6 +16,7 @@
     "concurrently": "^5.1.0"
   },
   "dependencies": {
+    "apollo-server-express": "^3.12.0",
     "axios": "^1.4.0",
     "bootstrap": "^5.3.1",
     "react-bootstrap": "^2.8.0"
